+++ Server/Core/Models/EntityPosition.cs
@@ -2,7 +2,7 @@ namespace BattleSimulator.Server.Models;
 
 public class EntityPosition
 {
-    public int x { get; set; }
-    public int y { get; set; }
+    public double x { get; set; }
+    public double y { get; set; }
     public string entityIdentifier { get; set; } = "";
 }
